Abstracted functions into a dedicated file
